Implemented (scrappy) article digest from wired.com. Read desc to hear me trash on it.
Ok so basically at the moment, the 'article getter and summariser' will call, get the top 5 articles from wired.com/rss and and then feed it into the one OAI llm. Ideally instead we feed it into a microllm to pick the most interesting title, then another small (but slightly bigger) llm to summarise it. This would be slightly slower but much cheaper. But hey. Do what works right.